requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.10.5
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Authentication and Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# LLM and AI libraries
langchain==0.3.25
langchain-core==0.3.74
langchain-openai==0.3.25
langchain-anthropic==0.3.19
langgraph==0.6.0
openai==1.86.0
anthropic==0.64.0

# PDF and document processing
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11

# LaTeX processing
pylatex==1.4.1

# CV/Resume generation
rendercv[full]==2.2

# Additional utilities
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.2
pyyaml==6.0.1 
supabase==2.3.4

# Email validation
email-validator==2.2.0